Refine collision detection
Before, the collision detection required the player to be within the
boundary to trigger actions (for example, the player has to partially
clip through a wall to reposition itself beside said wall).

Now the collision box is positioned to the position that it would expect
in the very next tick beforehand (collision box will still update at the
end of the current tick).